Final stuff done. Ready to publish. Go version will be abandoned.
