+++ server/handles.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	"github.com/Stridsvagn69420/Cyrkensia/utils"
-
 	"github.com/Stridsvagn69420/pringo"
 
 	"github.com/gofiber/fiber/v2"
