+++ utils/config.go
@@ -15,10 +15,44 @@ type ConfigType struct {
 	Icon     string `json:"hosticon"`
 	Uuid     string `json:"uuid"`
 	Locked   bool
-	Name     string `json:"vendorName"`
-	Access   string `json:"htpasswd"`
-	Pem      string `json:"pemTLS"`
-	Key      string `json:"keyTLS"`
+	Name     string  `json:"vendorName"`
+	Access   string  `json:"htpasswd"`
+	Pem      string  `json:"pemTLS"`
+	Key      string  `json:"keyTLS"`
+	Owners   []Owner `json:"owners"`
+}
+
+type Hostinfo struct {
+	Name      string  `json:"name"`
+	Hosticon  string  `json:"hosticon"`
+	Uuid      string  `json:"uuid"`
+	Secured   bool    `json:"secured"`
+	Size      int     `json:"size"`
+	Root      string  `json:"root"`
+	OriginURI string  `json:"originURI"`
+	Albums    []Album `json:"albums"`
+	Owners    []Owner `json:"owners"`
+}
+
+type Album struct {
+	Cover   string   `json:"cover"`
+	Dir     string   `json:"dir"`
+	Name    string   `json:"name"`
+	Files   []string `json:"files"`
+	Authors []Author `json:"authors"`
+}
+
+type Owner struct {
+	Name    string `json:"name"`
+	Website string `json:"website"`
+	Email   string `json:"email"`
+}
+
+type Author struct {
+	Name    string   `json:"name"`
+	Website string   `json:"website"`
+	Email   string   `json:"email"`
+	Songs   []string `json:"songs"`
 }
 
 var Config ConfigType
@@ -44,6 +78,7 @@ func LoadConfig(path string) {
 		assignIfEmpty(&Config.Access, tmpConfig.Access)
 		assignIfEmpty(&Config.Pem, tmpConfig.Pem)
 		assignIfEmpty(&Config.Key, tmpConfig.Key)
+		assignIfEmptyArray(&Config.Owners, tmpConfig.Owners)
 
 		// Special case for Locked and Port
 		if Config.Port == 0 {
@@ -60,5 +95,13 @@ func LoadConfig(path string) {
 func assignIfEmpty(variable *string, value string) {
 	if *variable == "" {
 		*variable = value
+	}
+}
+
+func assignIfEmptyArray(variable *[]Owner, value []Owner) {
+	if len(value) > 0 {
+		*variable = value
+	} else {
+		*variable = make([]Owner, 0)
 	}
 }
